Update to 'consul version' behaviour
After 24379e83df19c36f6bcbc607f391f11da7a90175 if consul was built
with just running "go build" the GitDescribe would be empty and consul
version would be empty.

This change alters the behaviour so that if consul is build without
proper ldflags the version will be postfixed with "dev" prerelease to
indicate that it is self compiled in a wrong way.

Should someone have a bug in such a binary at least devs should easily
see from the version number that binary has not been created by
recommended means.