Website: Delete the unnecessary backslash at the end shell command line
Shell thinks '\<newline>' as line continuation only when it's unquoted
or in double quotes. It's not necessary to manually mark '\<newline>'
in single quotes, because when shell sees the a single quote starts, it
continues the line automatically until it sees the matching single quote
ends.

Thus,
echo 'a\
  b'
would give the following output
a\
  b

And
echo 'a
  b'
gives the following output
a
  b

Since we don't want to leave backslashes in the resulting json files, we
should remove the line continuation backslashes inside single quotes.