Corrects formatting of JSON in snippet
The original snippet did not put the JSON output on its own line, which
made it look like it was part of the previous or next `curl` command.
This patch formats and highlights the command and output to appear as
they would in the terminal while also highlighting the multi-datacenter
elements of the output.